# optimized map::at