```cpp
template<int N, class L = brigand::range<int, 0, N>>
struct mk_map;

template<int N, class... T>
struct mk_map<N, brigand::list<T...>> {
  using type = brigand::map<brigand::pair<T,T>...>;
};

using list = brigand::at<mk_map<SIZE>::type, std::integral_constant<int, INDEX>>;
```

With INDEX = `0`, `SIZE/4`, `SIZE/2`, `SIZE-1` and `SIZE` (no differences).

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 34M | 00.06s - 35M | 00.07s - 36M | 00.09s - 39M | 00.14s - 46M
gcc | 00.04s - 19M | 00.05s - 21M | 00.06s - 23M | 00.09s - 28M | 00.16s - 43M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 34M | 00.06s - 35M | 00.06s - 35M | 00.07s - 36M | 00.08s - 38M
gcc | 00.04s - 19M | 00.05s - 20M | 00.05s - 21M | 00.07s - 22M | 00.10s - 26M